moved crossPromoSelector to the NavFrame.
