+++ src/app/components/NavFrame/index.jsx
@@ -1,5 +1,6 @@
 import './styles.less';
 
+import { some } from 'lodash/collection';
 import React from 'react';
 import { connect } from 'react-redux';
 
@@ -8,7 +9,46 @@ import InterstitialPromo from 'app/components/InterstitialPromo';
 import InterstitialListing from 'app/components/InterstitialListing';
 import EUCookieNotice from 'app/components/EUCookieNotice';
 import TopNav from 'app/components/TopNav';
-import { crossPromoSelector } from 'app/selectors/crossPromoSelector';
+import { flags as flagConstants } from 'app/constants';
+import featureFlags from 'app/featureFlags';
+
+const {
+  SMARTBANNER,
+  VARIANT_XPROMO_BASE,
+  VARIANT_XPROMO_LIST,
+  VARIANT_XPROMO_RATING,
+  VARIANT_XPROMO_LISTING,
+  VARIANT_XPROMO_SUBREDDIT,
+  VARIANT_XPROMO_CLICK,
+} = flagConstants;
+
+
+export function crossPromoSelector(state) {
+  const features = featureFlags.withContext({ state });
+
+  const showBanner = state.smartBanner.showBanner;
+  const showInterstitial = showBanner &&
+                           some([
+                             VARIANT_XPROMO_BASE,
+                             VARIANT_XPROMO_LIST,
+                             VARIANT_XPROMO_RATING,
+                             VARIANT_XPROMO_LISTING,
+                           ], variant => features.enabled(variant));
+  const showInterstitialListing = showBanner &&
+                                  some([
+                                    VARIANT_XPROMO_SUBREDDIT,
+                                  ], variant => features.enabled(variant));
+  const showSmartBanner = !showInterstitial && !showInterstitialListing && showBanner
+                       && !features.enabled(VARIANT_XPROMO_CLICK)
+                       && features.enabled(SMARTBANNER);
+
+  return {
+    showInterstitial,
+    showInterstitialListing,
+    showSmartBanner,
+  };
+}
+
 
 const NavFrame = props => {
   const {
